moved finder to doc package using store
